fix clang-tidy warnings

+++ src/timer.cpp
@@ -10,7 +10,7 @@ void Timer::start() { start_time = Clock::now(); }
 void Timer::end() {
   chrono::duration duration =
       chrono::duration_cast<chrono::nanoseconds>(Clock::now() - start_time);
-  double ms = duration.count() / 1e6;
+  double ms = static_cast<double>(duration.count()) / 1e6;
 
   if (ms >= 1000) {
     cout << "Done in " << ms / 1e3 << " s" << endl;
